+++ lib/modules/presentation/provider/dependency_injections.provider.dart
@@ -5,7 +5,7 @@ import '../../domain/usecases/delete_task.dart';
 import '../../domain/usecases/get_all_task.dart';
 import '../../domain/usecases/insert_task.dart';
 import '../../domain/usecases/update_task.dart';
-import '../controller/home_page_controller.dart';
+import '../controller/controller.dart';
 
 class DependecyInjectionsProvider extends StatelessWidget {
   final Widget? child;
@@ -34,7 +34,7 @@ class DependecyInjectionsProvider extends StatelessWidget {
               DeleteTaskUsecaseImpl(repository: context.read<TaskRepository>()),
         ),
         ChangeNotifierProvider(
-          create: (context) => HomePageController(
+          create: (context) => Controller(
             getAllTaskUsecase: context.read<GetAllTaskUsecase>(),
             insertTaskUsecase: context.read<InsertTaskUsecase>(),
             updateTaskUsecase: context.read<UpdateTaskUsecase>(),
